fix: improve init
